refactor: Rename MainWindow folder to Home
