+++ src/AStar.Dev.OneDrive.Sync.Client/Home/MainWindow.axaml
@@ -3,9 +3,9 @@
         xmlns:accounts="using:AStar.Dev.OneDrive.Sync.Client.Accounts"
         xmlns:syncronisation="using:AStar.Dev.OneDrive.Sync.Client.Syncronisation"
         xmlns:syncronisationConflicts="using:AStar.Dev.OneDrive.Sync.Client.SyncronisationConflicts"
-        xmlns:vm="using:AStar.Dev.OneDrive.Sync.Client.MainWindow"
+        xmlns:vm="using:AStar.Dev.OneDrive.Sync.Client.Home"
         Icon="avares://AStar.Dev.OneDrive.Sync.Client/Assets/astar.png"
-        x:Class="AStar.Dev.OneDrive.Sync.Client.MainWindow.MainWindow"
+        x:Class="AStar.Dev.OneDrive.Sync.Client.Home.MainWindow"
         x:DataType="vm:MainWindowViewModel"
         Title="AStar Dev - OneDrive Multi-Account Sync"
         Width="1200"
